CODSOFT1/requirements.txt
pandas
numpy
scikit-learn
scikit-surprise
